services/web/requirements.txt
attrs==19.3.0
bcrypt
email-validator==1.0.5
Flask==1.1.1
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
psycopg2-binary==2.8.4
requests==2.23.0
